[+FEATURE] Refactoring for publishing / part 4

+++ src/app/code/community/FireGento/AdminMonitoring/Model/Clean.php
@@ -31,30 +31,54 @@ class FireGento_AdminMonitoring_Model_Clean
     const XML_PATH_ADMINMONITORING_CLEAN_ENABLED = 'admin/firegento_adminmonitoring/enable_cleaning';
 
     /**
-     * Cleaning database-table
+     * Cronjob method for cleaning the database table.
      *
      * @return FireGento_AdminMonitoring_Model_Clean
      */
-    public function clean ()
+    public function scheduledCleanAdminMonitoring()
     {
-        if (!Mage::getStoreConfigFlag(self::XML_PATH_ADMINMONITORING_INTERVAL)) {
+        if (!Mage::getStoreConfigFlag(self::XML_PATH_ADMINMONITORING_CLEAN_ENABLED)) {
             return $this;
         }
 
-        if (Mage::getStoreConfig(self::XML_PATH_ADMINMONITORING_CLEAN_ENABLED)) {
-
-            $interval = Mage::getStoreConfig(self::XML_PATH_ADMINMONITORING_INTERVAL);
-
-            /** @var $adminMonitoringCollection FireGento_AdminMonitoring_Model_Resource_History_Collection */
-            $adminMonitoringCollection = Mage::getModel('firegento_adminmonitoring/history')
-                ->getCollection()
-                ->addFieldToFilter('created_at', array(
-                'lt' => new Zend_Db_Expr("DATE_SUB('" . now() . "', INTERVAL " . (int) $interval . " DAY)")));
-
-            foreach ($adminMonitoringCollection as $history) {
-                $history->delete();
-            }
+        try {
+            $this->clean();
+        } catch (Exception $e) {
+            Mage::logException($e);
         }
+
+        return $this;
+    }
+
+    /**
+     * Clean the database table for the given interval.
+     *
+     * @return FireGento_AdminMonitoring_Model_Clean
+     */
+    public function clean()
+    {
+        if (!Mage::getStoreConfig(self::XML_PATH_ADMINMONITORING_CLEAN_ENABLED)
+            || !Mage::getStoreConfigFlag(self::XML_PATH_ADMINMONITORING_INTERVAL)
+        ) {
+            return $this;
+        }
+
+        $interval = Mage::getStoreConfig(self::XML_PATH_ADMINMONITORING_INTERVAL);
+
+        /* @var $adminMonitoringCollection FireGento_AdminMonitoring_Model_Resource_History_Collection */
+        $adminMonitoringCollection = Mage::getModel('firegento_adminmonitoring/history')
+            ->getCollection()
+            ->addFieldToFilter(
+                'created_at',
+                array(
+                    'lt' => new Zend_Db_Expr("DATE_SUB('" . now() . "', INTERVAL " . (int) $interval . " DAY)")
+                )
+            );
+
+        foreach ($adminMonitoringCollection as $history) {
+            $history->delete();
+        }
+
         return $this;
     }
 }
